hardcoded the scheduling for demo
