+++ src/components/VetProfile/bookingSystem2.0/Booking.js
@@ -18,22 +18,22 @@ class Booking extends React.Component {
             // sat: props.week.saturday,
             // sun: props.week.sunday
         };
-        this.handleClick = this.handleClick.bind(this);
+        // this.handleClick = this.handleClick.bind(this);
     }
 
 
     handleClick= (e) => {
         // e.preventDefault();
         this.setState(prevState => ({ isFlipped: !prevState.isFlipped }));
-        this.setState({selectedDate: e})
-        if ((String(e).substring(0,3)) === "Mon") {this.setState({selectedWeekday: this.props.key1.monday})}
-        if ((String(e).substring(0,3)) === "Tue") {this.setState({selectedWeekday: this.props.key1.tuesday})}
-        if ((String(e).substring(0,3)) === "Wed") {this.setState({selectedWeekday: this.props.key1.wednesday})}
-        if ((String(e).substring(0,3)) === "Thu") {this.setState({selectedWeekday: this.props.key1.thursday})}
-        if ((String(e).substring(0,3)) === "Fri") {this.setState({selectedWeekday: this.props.key1.friday})}
-        if ((String(e).substring(0,3)) === "Sat") {this.setState({selectedWeekday: this.props.key1.saturday})}
-        if ((String(e).substring(0,3)) === "Sun") {this.setState({selectedWeekday: this.props.key1.sunday})}
-        console.log(this.props.key1)
+    //     this.setState({selectedDate: e})
+    //     if ((String(e).substring(0,3)) === "Mon") {this.setState({selectedWeekday: this.props.key1[1]})}
+    //     if ((String(e).substring(0,3)) === "Tue") {this.setState({selectedWeekday: this.props.key1[2]})}
+    //     if ((String(e).substring(0,3)) === "Wed") {this.setState({selectedWeekday: this.props.key1[3]})}
+    //     if ((String(e).substring(0,3)) === "Thu") {this.setState({selectedWeekday: this.props.key1[4]})}
+    //     if ((String(e).substring(0,3)) === "Fri") {this.setState({selectedWeekday: this.props.key1[5]})}
+    //     if ((String(e).substring(0,3)) === "Sat") {this.setState({selectedWeekday: this.props.key1[6]})}
+    //     if ((String(e).substring(0,3)) === "Sun") {this.setState({selectedWeekday: this.props.key1[7]})}
+    //     console.log(this.props.key1)
     }
 
     handleTimeSelection(e, item){
@@ -43,7 +43,7 @@ class Booking extends React.Component {
 
     render() {
         return (
-            <ReactCardFlip isFlipped={this.state.isFlipped} flipDirection="vertical">
+            <ReactCardFlip isFlipped={this.state.isFlipped} flipDirection="horizontal">
                 <MuiPickersUtilsProvider utils={DateFnsUtils}>
                     <DatePicker style={{textAlign: 'center'}}
                         autoOk
@@ -55,14 +55,17 @@ class Booking extends React.Component {
                     />
                 </MuiPickersUtilsProvider>
 
-                <div>
-                    <button onClick={this.handleClick}>Click to flip Back</button>
-                    <ul>
-                        {this.state.selectedWeekday.map((item) => (
-                            <button id="submitButton" onClick={(e, item) => this.handleTimeSelection(e, item)}>item</button>
-                            // <ItemBlock key3={this.selectedDate} key1={item}/>
-                        ))}
-                    </ul>
+                <div style={{textAlign: 'center'}}>
+                    <button onClick={this.handleClick}>Return to Calendar</button><br/>
+                    <button id="submitButton">6:00pm</button><br/>
+                    <button id="submitButton" onClick={(e, item) => this.handleTimeSelection(e, item)}>7h00pm</button><br/>
+                    <button id="submitButton" onClick={(e, item) => this.handleTimeSelection(e, item)}>8h00pm</button><br/>
+                    {/*<ul>*/}
+                    {/*    {this.state.selectedWeekday.map((item) => (*/}
+                    {/*        <button id="submitButton" onClick={(e, item) => this.handleTimeSelection(e, item)}>item</button>*/}
+                    {/*        // <ItemBlock key3={this.selectedDate} key1={item}/>*/}
+                    {/*    ))}*/}
+                    {/*</ul>*/}
                 </div>
             </ReactCardFlip>
         )
